Updated BSH/HSB attention layouts to use unified logic
GitOrigin-RevId: 03bdaf947c48ebf4b216eb7183d6cc388e010f10
