+++ src/transformers_neuronx/llama/hlo.py
@@ -15,7 +15,7 @@
 from typing import Optional
 
 from transformers_neuronx import hlo
-from transformers_neuronx.layers import transformer, rotary, generation
+from transformers_neuronx.layers import transformer, rotary, attention, attention_utils
 from transformers_neuronx.llama.config import LlamaConfig
 from transformers_neuronx.config import NeuronConfig
 from transformers_neuronx.constants import LAYOUT_BSH, LAYOUT_HSB
@@ -63,7 +63,6 @@ class LlamaForSamplingNoEmbeddingHlo:
         )
         mask, active_mask = hlo.attention_mask(cache_ids, start_ids, self.n_positions)
         return hidden, last_token_id, pos_embed, cache_ids, start_ids, mask, active_mask
-
 
     def layer(
             self, hidden, last_token_id, pos_embed, cache_ids, start_ids, mask, active_mask,
@@ -125,12 +124,9 @@ class LlamaForSamplingNoEmbeddingHlo:
     ):
         d_head = self.config.attention_head_size
         tp_degree = self.config.tp_degree
-
-        if self.neuron_config and self.neuron_config.attention_layout == LAYOUT_BSH:
-            import transformers_neuronx.layers.attention as attention
-        else:
-            import transformers_neuronx.layers.attention_hsb as attention
-        from transformers_neuronx.layers import attention_utils
+        n_kv_heads_tp = self.config.num_key_value_heads
+        if n_kv_heads_tp is not None:
+            n_kv_heads_tp = n_kv_heads_tp // tp_degree
 
         # Q = (hidden @ wQ) + bQ
         # K = (hidden @ wK) + bK
@@ -144,6 +140,7 @@ class LlamaForSamplingNoEmbeddingHlo:
             neuron_config=self.neuron_config,
             tp_degree=tp_degree,  # TODO: include tp_degree into neuron_config
             shard_over_batch=self.shard_over_batch,
+            n_kv_heads_tp=n_kv_heads_tp,
         )
 
         # Q = Rotate(Q)
